+++ src/wrapper/preact/test.ts
@@ -48,7 +48,7 @@ describe('Preact wrapper', () => {
     )
     const el = getByText('Button text')
     expect('className' in el && el.className).toEqual(
-      'Button-active-xxx Button-color-red-xxx Button-xxx'
+      'Button-xxx Button-active-xxx Button-color-red-xxx'
     )
   })
 })